chore: move events to a single class
