+++ common/src/main/java/com/wynntils/features/ui/CustomLoadingScreenFeature.java
@@ -10,7 +10,7 @@ import com.wynntils.core.persisted.config.ConfigCategory;
 import com.wynntils.mc.event.LoadingProgressEvent;
 import com.wynntils.mc.event.LocalSoundEvent;
 import com.wynntils.mc.event.ScreenOpenedEvent;
-import com.wynntils.mc.event.ServerResourcePackLoadEvent;
+import com.wynntils.mc.event.ServerResourcePackEvent;
 import com.wynntils.mc.event.SubtitleSetTextEvent;
 import com.wynntils.mc.event.TickAlwaysEvent;
 import com.wynntils.mc.event.TitleSetTextEvent;
@@ -65,7 +65,7 @@ public class CustomLoadingScreenFeature extends Feature {
     }
 
     @SubscribeEvent
-    public void onResourcePack(ServerResourcePackLoadEvent e) {
+    public void onResourcePack(ServerResourcePackEvent.ServerResourcePackLoadEvent e) {
         if (loadingScreen == null) return;
 
         loadingScreen.setMessage(I18n.get("feature.wynntils.customLoadingScreen.resourcePack"));
